A function that reads a text file and prints it to the POSIX standard output

+++ 0x15-file_io/0-read_textfile.c
@@ -31,7 +31,7 @@ ssize_t read_textfile(const char *filename, size_t letters)
 		return (0);
 	}
 	bytes_written = write(STDOUT_FILENO, buffer, bytes_read);
-	if (bytes_written == -1 || bytes_written != bytes_read)
+	if (bytes_written == -1 || (size_t)bytes_written != (size_t)bytes_read)
 	{
 		free(buffer);
 		close(file_descripter);
